refactor(utils): consolidate string/timestamp formatters into stringUtils
src/shared/utils/string.ts duplicated the "single home for generic string
helpers" that AGENTS.md documents for @utils/text/stringUtils, giving
formatRelativeTime/formatShortDateTime/formatBytes a second, undocumented
home in a different alias namespace with no browser-safety tripwire. Move
the three functions into stringUtils.ts, delete the shared/utils file, and
repoint its 9 call sites (CLI, extension frontends, tools) plus fold the
one-off DateTimeFormatters.vitest.ts suite into the existing
StringUtils.vitest.ts.
