--- src/shared/utils/string.ts
@@ -1,36 +1,0 @@
-import { intlFormatDistance } from 'date-fns';
-import prettyBytes from 'pretty-bytes';
-
-const DATE_TIME_FORMATTER = new Intl.DateTimeFormat(undefined, {
-  year: 'numeric',
-  month: 'short',
-  day: 'numeric',
-  hour: '2-digit',
-  minute: '2-digit',
-});
-
-/** Calendar-aware "X ago" / "in X" via Intl.RelativeTimeFormat (handles future timestamps too). */
-export function formatRelativeTime(timestamp: number): string {
-  if (!timestamp) return '';
-  return intlFormatDistance(timestamp, Date.now());
-}
-
-/**
- * Formats a date-parsable value as a compact, locale-aware absolute
- * timestamp (short month, no seconds) — the shared "list-item timestamp"
- * shape used across the settingsView History and Memory lists. Returns
- * `null` for missing/invalid input so callers can supply their own fallback
- * copy (e.g. an "Updated: unknown" meta-strip entry).
- */
-export function formatShortDateTime(
-  value: string | number | Date | null | undefined,
-): string | null {
-  if (!value) return null;
-  const date = new Date(value);
-  if (Number.isNaN(date.getTime())) return null;
-  return DATE_TIME_FORMATTER.format(date);
-}
-
-export function formatBytes(bytes: number): string {
-  return prettyBytes(bytes, { binary: true });
-}
